refactor the common functionality of GeoJSON  and ShapeFile Reader
